refactor: rename unsplit to merge and try_unsplit to unsplit
try_unsplit only has one fail case that's side-effect free. It doesn't pose a danger to public users. Why not make it pub?

There's two use cases where I will want to merge Bytes

   1. merge two non-contiguous Bytes to have a convenient single view
   2. merge two contiguous Bytes that have previously been split off

In the first case, if the Bytes happen to be contiguous, then the O(1) operation is just a bonus. But in the second case, I expect that the Bytes must be contiguous. Otherwise it would be an error to merge them. unsplit hides this failure and is general over both cases. Maybe rename unsplit to merge and keep as is, but rename try_unsplit to unsplit and keep as is and expose both.